Introduced ability to post a draft review.
A draft review can be posted to Review Board using the post subcommand
provided by the scripts rbt and rbt.py. This implementation is
primitive.  The file, summary and related review information is not
populated.

Corrected a problem introduced in the previous commit that required the
Review Board URL to include the path to the REST API.